requirements.txt
certifi==2018.1.18
cffi==1.11.4
cycler==0.10.0
decorator==4.0.11
imageio==2.1.2
matplotlib==2.1.2
moviepy==0.2.3.2
numpy==1.14.0
olefile==0.45.1
opencv-python==3.4.0.12
Pillow==5.0.0
pycparser==2.18
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2018.3
PyYAML==3.12
scikit-learn==0.19.1
scipy==1.0.1
six==1.11.0
sklearn==0.0
torch==0.3.1.post2
torchvision==0.2.0
tqdm==4.11.2
certifi==2018.1.18
cffi==1.11.4
cycler==0.10.0
decorator==4.0.11
imageio==2.1.2
matplotlib==2.1.2
moviepy==0.2.3.2
numpy==1.14.0
olefile==0.45.1
opencv-python==3.4.0.12
Pillow==5.0.0
pycparser==2.18
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2018.3
PyYAML==3.12
scikit-learn==0.19.1
scipy==1.0.1
six==1.11.0
sklearn==0.0
torch==0.3.1.post2
torchvision==0.2.0
tqdm==4.11.2
